ref: Move Snuba specific functionality out of `snuba.utils.streams.backends`
This is an attempt to separate the Snuba specific parts of Kafka configuration
(specifically the part that depends on Snuba's settings formats and topics)
from the parts that are generic and can be shared with other codebases such as
Sentry.
