+++ snuba/cli/multistorage_consumer.py
@@ -11,9 +11,8 @@ from snuba.datasets.storages import StorageKey
 from snuba.datasets.storages.factory import WRITABLE_STORAGES, get_writable_storage
 from snuba.environment import setup_logging, setup_sentry
 from snuba.utils.metrics.wrapper import MetricsWrapper
-from snuba.utils.streams.backends.kafka import (
-    KafkaConsumer,
-    KafkaConsumerWithCommitLog,
+from snuba.utils.streams.backends.kafka import KafkaConsumer, KafkaConsumerWithCommitLog
+from snuba.utils.streams.configuration_builder import (
     build_kafka_consumer_configuration,
     build_kafka_producer_configuration,
 )
